Move `LocalHackrStoryLoader` and collaborators  to production
