+++ HackrNewsTests/Hackr News Feed Cache/Story/Helpers/HackrNewsStoryStoreSpy.swift
@@ -1,0 +1,44 @@
+//
+//  Copyright © 2021 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import Foundation
+import HackrNews
+
+class HackrNewsStoryStoreSpy: HackrNewsStoryStore {
+    enum Message: Equatable {
+        case deletion(Story)
+        case insertion(Story, Date)
+    }
+
+    private(set) var receivedMessages = [Message]()
+    private(set) var deletionCompletions = [DeletionCompletion]()
+
+    func delete(_ story: Story, completion: @escaping DeletionCompletion) {
+        deletionCompletions.append(completion)
+        receivedMessages.append(.deletion(story))
+    }
+
+    func completeDeletion(with error: Error, at index: Int = 0) {
+        deletionCompletions[index](.failure(error))
+    }
+
+    func completeDeletionSuccessfully(at index: Int = 0) {
+        deletionCompletions[index](.success(()))
+    }
+
+    private(set) var insertionCompletions = [InsertionCompletion]()
+
+    func insert(story: Story, with timestamp: Date, completion: @escaping InsertionCompletion) {
+        insertionCompletions.append(completion)
+        receivedMessages.append(.insertion(story, timestamp))
+    }
+
+    func completeInsertion(with error: Error, at index: Int = 0) {
+        insertionCompletions[index](.failure(error))
+    }
+
+    func completeInsertionSuccessfully(at index: Int = 0) {
+        insertionCompletions[index](.success(()))
+    }
+}
